backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "tech-admin-backend"
version = "0.1.0"
description = "热流道技术管理端 — P0/M0 后端 API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.1",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.18",
    "email-validator>=2.2.0",
    "reportlab>=4.4.10",
]

[project.scripts]
serve = "app.serve_cli:main"

[dependency-groups]
dev = [
    "ruff>=0.8.4",
    "httpx>=0.28.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
